refactor(QoL): :recycle: componentize Home
